chore(voyager): cleanup
- remove $ from memo
- add a log when fetching account infoo
- lazily decode evm errors
- remove invalid private keys from voyager config (lol)
- update stargaze flake input
