+++ lib/chain-utils/src/cosmos_sdk.rs
@@ -113,6 +113,7 @@ pub trait CosmosSdkChainExt: CosmosSdkChain {
     }
 
     async fn account_info(&self, account: &str) -> BaseAccount {
+        tracing::debug!(%account, "fetching account");
         let Any(account) =
             protos::cosmos::auth::v1beta1::query_client::QueryClient::connect(self.grpc_url())
                 .await
@@ -146,7 +147,7 @@ pub trait CosmosSdkChainExt: CosmosSdkChain {
             body_bytes: tx::v1beta1::TxBody {
                 messages: messages.clone().into_iter().collect(),
                 // TODO(benluelo): What do we want to use as our memo?
-                memo: format!("Voyager ${}", env!("CARGO_PKG_VERSION")),
+                memo: format!("Voyager {}", env!("CARGO_PKG_VERSION")),
                 timeout_height: 123_123_123,
                 extension_options: vec![],
                 non_critical_extension_options: vec![],
